Remove utils_test, use simple naming instead of gen_random_key()
